@@ -4,67 +4,54 @@
   "description": "A standalone application for ingesting content onto the Eluvio fabric.",
   "main": "index.js",
   "author": "Eluvio, Inc.",
+  "type": "module",
   "license": "MIT",
   "scripts": {
-    "serve": "webpack-dev-server --hot --host 0.0.0.0",
-    "build": "rm -rf dist/* && webpack-cli --mode production --no-devtool",
-    "lint": "npm run fix-scss ; npm run lint-scss ; npm run fix-js",
-    "fix-scss": "sass-lint-auto-fix -c .scss-lint.yml",
-    "lint-scss": "stylelint \"src/**/*.scss\"",
-    "lint-js": "./node_modules/.bin/eslint \"src/**/*.js\"",
-    "fix-js": "npm run lint-js -- --fix",
+    "serve": "rm -rf node_modules/.vite; vite",
+    "build": "rm -rf dist/* ; vite build",
+    "preview": "vite preview",
+    "lint": "npm-run-all -s -c lint-js-fix lint-scss-fix",
+    "lint-js": "eslint ./src --ext js,jsx --report-unused-disable-directives",
+    "lint-js-fix": "eslint ./src --fix --ext js,jsx --report-unused-disable-directives",
+    "lint-scss": "npx stylelint \"src/**/*.{css,scss}\"",
+    "lint-scss-fix": "npx stylelint \"src/**/*.{css,scss}\" --fix",
     "prepare": "husky install"
   },
   "dependencies": {
     "@eluvio/elv-abr-profile": "^1.0.1",
-    "@eluvio/elv-client-js": "^4.0.72",
+    "@eluvio/elv-client-js": "^4.0.93",
     "@eluvio/elv-lro-status": "^3.0.7",
+    "@mantine/core": "^7.11.2",
     "@radix-ui/react-dialog": "^1.0.0",
     "@radix-ui/react-tooltip": "^1.0.0",
-    "mobx": "^6.6.0",
-    "mobx-react": "^7.5.0",
+    "mobx": "^6.13.1",
+    "mobx-react-lite": "^4.0.7",
     "pretty-bytes": "^6.0.0",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "17.0.2",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.0",
-    "react-inlinesvg": "^3.0.1",
-    "react-router-dom": "5.2.0",
-    "url-join": "^5.0.0"
+    "react-router-dom": "^6.3.0",
+    "url-join": "^5.0.0",
+    "vite-plugin-commonjs": "^0.10.1",
+    "vite-plugin-static-copy": "^1.0.6"
   },
   "devDependencies": {
     "@babel/core": "^7.18.0",
     "@babel/eslint-parser": "^7.18.2",
     "@babel/preset-env": "^7.18.0",
     "@babel/preset-react": "^7.17.12",
-    "babel-loader": "^8.2.5",
-    "copy-webpack-plugin": "^11.0.0",
-    "crypto-browserify": "^3.12.0",
-    "css-loader": "^6.7.1",
+    "@vitejs/plugin-react-swc": "^3.6.0",
     "eslint": "7.32.0",
-    "eslint-plugin-react": "^7.30.0",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
+    "eslint-plugin-react": "^7.34.4",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.8",
     "husky": "^8.0.1",
-    "image-webpack-loader": "^8.1.0",
-    "process": "^0.11.10",
+    "npm-run-all": "^4.1.5",
     "sass": "^1.52.1",
-    "sass-lint-auto-fix": "^0.21.2",
     "sass-loader": "^13.0.0",
-    "stream-browserify": "^3.0.0",
-    "style-loader": "^3.3.1",
     "stylelint": "^14.8.5",
     "stylelint-config-standard-scss": "^4.0.0",
-    "svg-inline-loader": "^0.8.2",
-    "vm-browserify": "^1.1.2",
-    "webpack": "5.72.1",
-    "webpack-cli": "4.10.0",
-    "webpack-dev-server": "4.9.0"
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env",
-      "@babel/preset-react"
-    ]
+    "vite": "^5.3.4"
   }
 }
