@@ -1,6 +1,6 @@
 import {configure, flow, makeObservable, observable} from "mobx";
 import {FrameClient} from "@eluvio/elv-client-js/src/FrameClient";
-import IngestStore from "Stores/IngestStore";
+import IngestStore from "@/stores/IngestStore";
 
 // Force strict mode so mutations are only allowed within actions.
 configure({
@@ -33,8 +33,10 @@ class RootStore {
 
       this.networkInfo = yield this.client.NetworkInfo();
     } catch(error) {
+      /* eslint-disable no-console */
       console.error("Failed to initialize application");
       console.error(error);
+      /* eslint-enable no-console */
     } finally {
       this.loaded = true;
     }
@@ -44,6 +46,7 @@ class RootStore {
     try {
       return this.client.utils.FromB64(string);
     } catch(error) {
+      // eslint-disable-next-line no-console
       console.error(`Unable to decode ${string}.`, error);
     }
   };
